WebView2 Part 8

+++ Sources/Libraries/HTMLViewer.ahk
@@ -1944,16 +1944,47 @@ class WebView2 extends WebView2.Base {
 
 class HTMLViewer {
 	iControl := false
-	iWebView2 := false
-
-	iHTML := false
-	iHTMLFile := false
 
 	Control {
 		Get {
 			return this.iControl
 		}
 	}
+
+	Document {
+	}
+
+	Location {
+	}
+
+	__New(control) {
+		this.iControl := control
+	}
+
+	Resized() {
+	}
+
+	Navigate(uri) {
+	}
+
+	Open() {
+	}
+
+	Close() {
+	}
+
+	Write(html, force := false) {
+	}
+
+	Reload() {
+	}
+}
+
+class WebView2Viewer {
+	iWebView2 := false
+
+	iHTML := false
+	iHTMLFile := false
 
 	WebView2 {
 		Get {
@@ -1977,9 +2008,9 @@ class HTMLViewer {
 		local application := Strsplit(A_ScriptName, ".")[1]
 		local rnd := Random(1, 1000)
 
-		this.iControl := control
-
 		this.iHTMLFile := temporaryFileName("HTML\" . application . "\" . "html" . rnd, "html")
+
+		super.__New(control)
 	}
 
 	Show() {
@@ -1999,15 +2030,6 @@ class HTMLViewer {
 	Resized() {
 		if this.WebView2
 			this.WebView2.Fill()
-	}
-
-	Navigate(uri) {
-	}
-
-	Open() {
-	}
-
-	Close() {
 	}
 
 	Write(html, force := false) {
@@ -2030,6 +2052,58 @@ class HTMLViewer {
 	}
 }
 
+class IEViewer extends HTMLViewer {
+	iExplorer := false
+	iControl := false
+
+	Control {
+		Get {
+			return this.iControl
+		}
+	}
+
+	Explorer {
+		Get {
+			return this.iExplorer
+		}
+	}
+
+	Document {
+		Get {
+			return this.Explorer.document
+		}
+	}
+
+	Location {
+		Get {
+			return this.Explorer.location
+		}
+	}
+
+	__New(control) {
+		this.iExplorer := control.Value
+	}
+
+	Navigate(uri) {
+	}
+
+	Open() {
+		this.Document.open()
+	}
+
+	Close() {
+		this.Document.close()
+	}
+
+	Write(html, force := false) {
+		this.Document.write(html)
+	}
+
+	Reload() {
+		this.Location.reload()
+	}
+}
+
 
 ;;;-------------------------------------------------------------------------;;;
 ;;;                        Private Function Section                         ;;;
@@ -2044,9 +2118,9 @@ CoTaskMem_String(ptr) {
 }
 
 initializeHTMLViewer() {
-	createHTMLViewer(window, arguments*) {
+	createWebView2Viewer(window, arguments*) {
 		local control := window.Add("Picture", arguments*)
-		local viewer := HTMLViewer(control)
+		local viewer := WebView2Viewer(control)
 
 		control.HTMLViewer := viewer
 		control.Document := viewer
@@ -2087,11 +2161,51 @@ initializeHTMLViewer() {
 		return control
 	}
 
-	deleteDirectory(kTempDirectory . "HTML\" . Strsplit(A_ScriptName, ".")[1])
+	createIEViewer(window, options) {
+		local control := window.Add("ActiveX", options, "shell.explorer")
+		local viewer := IEViewer(control)
+		local explorer := viewer.Explorer
 
-	DirCreate(kTempDirectory . "HTML\" . Strsplit(A_ScriptName, ".")[1])
+		explorer.navigate("about:blank")
 
-	Window.DefineCustomControl("HTMLViewer", createHTMLViewer)
+		control.HTMLViewer := viewer
+		control.Document := explorer.document
+
+		show() {
+			control.Visible := true
+		}
+
+		hide() {
+			control.Visible := false
+		}
+
+		width() {
+			return explorer.Width
+		}
+
+		height() {
+			return explorer.Height
+		}
+
+		control.Show := (*) => show()
+		control.Hide := (*) => hide()
+		control.Navigate := (obj, uri) => explorer.Navigate(uri)
+		control.Resized := (*) => viewer.Resized()
+		control.getWidth := (*) => width()
+		control.getHeight := (*) => height()
+
+		return control
+	}
+
+	if false {
+		deleteDirectory(kTempDirectory . "HTML\" . Strsplit(A_ScriptName, ".")[1])
+
+		DirCreate(kTempDirectory . "HTML\" . Strsplit(A_ScriptName, ".")[1])
+
+		Window.DefineCustomControl("HTMLViewer", createWebView2Viewer)
+	}
+	else
+		Window.DefineCustomControl("HTMLViewer", createIEViewer)
 }
